consolidate python routines into less files; update documentation to mention python prefida
