+++ lib/python/fidasim/__init__.py
@@ -1,0 +1,1 @@
+from fidasim.preprocessing import prefida
